Play the Beetle secondary animation shake
The secondary idle animation shake was not playing. Though there were code that seams to have intended the animation to play at random intervals.

Ensure that the timer running and when Idle starting the shake animation by traveling the state machine to the Shake state. Removed the continuous call to Idle that cased immediate travel out of Shake again and instead called idle when the Beetle loses track of the player.