fix: fix gamepad bindings (#25)
The camera controls were connected to the wrong axes, and the pause action
had no mapping so it was impossible to start the game.

Also made it autodetect when a controller is connected, so that it shows
the controller mappings by default (otherwise there's no way to toggle the
button currently).

And changed the "Pause" label to "Pause / Resume" to make it clearer that
it's the button that will start the game.

New bindings tested with an Xbox Series controller on Windows 10 and Linux.