Convert to Godot 4.1, breaks compat with Godot 4.0
Godot 4.1 sadly includes a compat breaking change in the way .glb and .blend files
with import hints like `-convcol` are imported, when they contain a trailing `.001`
number generated by Blender.

Godot doesn't support `.` in node names, and in 4.0 it would simply remove the
invalid characters. In 4.1+, it replaces invalid characters with `_`.

This creates a clash between the names of objects inherited from a .glb scene
with local modifications in the Godot scene, leading to a lot of errors.
I solved it by replacing the names manually in the scene files to match what
Godot 4.1 will generate.

The `res://.godot` folder generated with Godot 4.0 should be removed to force a
clean reimport with Godot 4.1.