Remove binary data from grenade.tscn
Had to disable "Reset On Save" on the AnimationPlayer as it would keep
changing the `grenade.material` file every time a new session is saved.

There's a Godot bug to investigate there, as there should be no change
to the material when just resetting the value which is already the default.